update HttpSequencRangeGetter and DefaultDDALDataSource
