+++ ddal-sequence/src/main/java/org/hellojavaer/ddal/sequence/HttpSequencRangeGetter.java
@@ -15,9 +15,12 @@
  */
 package org.hellojavaer.ddal.sequence;
 
-import org.hellojavaer.ddal.core.utils.HttpUtils;
 import org.hellojavaer.ddal.sequence.exception.GetSequenceFailedException;
 
+import java.io.*;
+import java.net.HttpURLConnection;
+import java.net.URL;
+import java.net.URLEncoder;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -118,5 +121,80 @@ public class HttpSequencRangeGetter implements SequenceRangeGetter {
         sequenceRange.setBeginValue(beginValue);
         sequenceRange.setEndValue(endValue);
         return sequenceRange;
+    }
+
+    static class HttpUtils {
+
+        private static final String USER_AGENT = "Mozilla/5.0";
+
+        public static String sendPost(String url, Map<String, Object> params) {
+            try {
+                URL obj = new URL(url);
+                HttpURLConnection con = (HttpURLConnection) obj.openConnection();
+                con.setRequestMethod("POST");
+                con.setRequestProperty("User-Agent", USER_AGENT);
+                con.setRequestProperty("Accept-Language", "en-US,en;q=0.5");
+                con.setDoOutput(true);
+                // send
+                DataOutputStream wr = null;
+                try {
+                    wr = new DataOutputStream(con.getOutputStream());
+                    if (params != null && !params.isEmpty()) {
+                        StringBuilder sb = new StringBuilder();
+                        for (Map.Entry<String, Object> entry : params.entrySet()) {
+                            sb.append(URLEncoder.encode(entry.getKey(), "UTF-8"));
+                            sb.append('=');
+                            sb.append(URLEncoder.encode(entry.getValue().toString(), "UTF-8"));
+                            sb.append('&');
+                        }
+                        sb.deleteCharAt(sb.length() - 1);
+                        wr.write(sb.toString().getBytes("UTF-8"));
+                    }
+                    wr.flush();
+                } finally {
+                    closeIO(wr);
+                }
+                // get
+                BufferedReader br = null;
+                InputStream in;
+                if (con.getResponseCode() < HttpURLConnection.HTTP_BAD_REQUEST) {
+                    in = con.getInputStream();
+                } else {
+                    in = con.getErrorStream();
+                }
+                try {
+                    br = new BufferedReader(new InputStreamReader(in));
+                    String inputLine;
+                    StringBuilder response = new StringBuilder();
+                    while ((inputLine = br.readLine()) != null) {
+                        response.append(inputLine);
+                        response.append('\n');
+                    }
+                    int responseCode = con.getResponseCode();
+                    if (responseCode != 200) {
+                        throw new IllegalStateException("http code " + responseCode + "\n" + response.toString());
+                    } else {
+                        return response.toString();
+                    }
+                } finally {
+                    closeIO(br);
+                }
+            } catch (Exception e) {
+                if (e instanceof RuntimeException) {
+                    throw (RuntimeException) e;
+                } else {
+                    throw new RuntimeException(e);
+                }
+            }
+        }
+
+        private static void closeIO(Closeable closeable) {
+            if (closeable != null) {
+                try {
+                    closeable.close();
+                } catch (IOException e) {
+                }
+            }
+        }
     }
 }
